+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
   * The ability to print a short help message, better suited to sub-commands.
+  * Autocompletion support using argcomplete. A new dependency of the library.
 
 ### Changed